+++ tests/Jroc.Tests/Node/Console/ExecutionTests.cs
@@ -9,5 +9,9 @@ namespace Jroc.Tests.Node.Console
         [Fact]
         public Task Console_Undici_Transform_Table()
             => ExecutionTest(nameof(Console_Undici_Transform_Table));
+
+        [Fact]
+        public Task Require_NodeConsole_Log_Variadic()
+            => ExecutionTest(nameof(Require_NodeConsole_Log_Variadic));
     }
 }