+++ tests/Jroc.Tests/Node/Path/JavaScript/Require_Path_UpdateOverride.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const path = require("path");
+
+path.join++;
+
+try {
+    path.join("a", "b");
+} catch (error) {
+    console.log(error.name);
+}